dcache-axi-fix

+++ MemSubSystem/include/PeripheralAxi.h
@@ -31,7 +31,7 @@ struct PeripheralAxiWriteOut {
   uint32_t req_addr = 0;
   uint8_t req_total_size = 0;
   uint8_t req_id = 0;
-  uint32_t req_wstrb = 0;
+  uint64_t req_wstrb = 0;
   uint32_t req_wdata[DCACHE_LINE_WORDS] = {};
   bool resp_ready = false;
 };
@@ -76,6 +76,6 @@ public:
 private:
   static uint8_t access_size_bytes(uint8_t func3);
   static uint32_t pack_write_data(uint32_t data, uint8_t func3);
-  static uint32_t pack_write_strb(uint8_t func3);
+  static uint64_t pack_write_strb(uint8_t func3);
   static uint32_t extract_read_data(uint32_t raw, uint8_t func3);
 };
